Implementacija kontrola

+++ Project/HospitalInformationSystem/Windows/RoomManagementWindow.xaml.cs
@@ -1,5 +1,6 @@
 ﻿using System.Windows;
 using BusinessLogic;
+using Model;
 
 namespace HospitalInformationSystem.Windows
 {
@@ -11,6 +12,8 @@ namespace HospitalInformationSystem.Windows
         public RoomManagementWindow()
         {
             InitializeComponent();
+            //confirmButton.IsEnabled = false;
+            //tryToEnableConfirmButton();
         }
 
         private void confirmButton_Click(object sender, RoutedEventArgs e)
@@ -20,23 +23,33 @@ namespace HospitalInformationSystem.Windows
             OneRoomWindow oneRoomWindow = new OneRoomWindow();
             DeleteOneRoomWindow deleteOneRoomWindow = new DeleteOneRoomWindow();
 
-
-            if ((bool)newRadioButton.IsChecked)
-                newRoomWindow.Show();
-            else if ((bool)allRoomsRadioButton.IsChecked)
-                allRoomsWindow.Show();
-            else if ((bool)oneRoomRadioButton.IsChecked)
-                oneRoomWindow.Show();
-            else if ((bool)deleteAllRoomsRadioButton.IsChecked)
+            if (RoomDataBase.getInstance().GetRoom().Count != 0)
             {
-                deleteAllRooms();
-                MessageBox.Show("Sve prostorije su sada obrisane iz sistema.", "Operacija brisanja", MessageBoxButton.OK, MessageBoxImage.Information);
+                if ((bool)allRoomsRadioButton.IsChecked)
+                    allRoomsWindow.Show();
+                else if ((bool)oneRoomRadioButton.IsChecked)
+                    oneRoomWindow.Show();
+                else if ((bool)deleteAllRoomsRadioButton.IsChecked)
+                {
+                    deleteAllRooms();
+                    MessageBox.Show("Sve prostorije su sada obrisane iz sistema.", "Operacija brisanja", MessageBoxButton.OK, MessageBoxImage.Information);
+                }
+                else if ((bool)deleteOneRoomRadioButton.IsChecked)
+                    deleteOneRoomWindow.Show();
             }
-            else if ((bool)deleteOneRoomRadioButton.IsChecked)
-                deleteOneRoomWindow.Show();
+            else if ((bool)newRadioButton.IsChecked)
+                newRoomWindow.Show();
+            else
+                MessageBox.Show("Ne možete pristupiti odabranoj funkcionalnosti jer u sistemu ne postoji nijedna prostorija.", "Prazna baza", MessageBoxButton.OK, MessageBoxImage.Warning);
 
               
 
+        }
+
+        private void tryToEnableConfirmButton()
+        {
+            if ((bool)allRoomsRadioButton.IsChecked || (bool)oneRoomRadioButton.IsChecked || (bool)deleteOneRoomRadioButton.IsChecked || (bool)deleteAllRoomsRadioButton.IsChecked || (bool)newRadioButton.IsChecked)
+                confirmButton.IsEnabled = true;
         }
 
         private void deleteAllRooms()
@@ -45,5 +58,9 @@ namespace HospitalInformationSystem.Windows
 
             management.DeleteAllRooms();
         }
+
+      
+
+       
     }
 }
